Some hotfixes

+++ manage.php
@@ -8,28 +8,24 @@ $count = count($sequence);
 $last = $count - 1;
 if ($mode == 'init') {
     foreach ($sequence as $key=>$value) {
-        if (!file_exists($value)) {
-            if (strpos($value, ':') !== false) {
-                $valSep = explode(':', $value);
-                $valName = $valSep[0];
-                $valType = $valSep[1];
+        if (strpos($value, ':') !== false) {
+            $valSep = explode(':', $value);
+            $valName = $valSep[0];
+            $valType = $valSep[1];
+            if (!file_exists($valName)) {
+                mkdir($valName);
+                chmod($valName, 0777);
                 if ($valType == 'profile') {
-                    mkdir($valName);
-                    chmod($valName, 0777);
                     file_put_contents($valName.'/rating', 0);
                     chmod($valName.'/rating', 0777);
                     file_put_contents($valName.'/mode', 0);
                     chmod($valName.'/mode', 0777);
                 } elseif ($valType == 'system') {
-                    mkdir($valName);
-                    chmod($valName, 0777);
                     if (file_exists('get.php')) {
                         copy('get.php', $valName.'/get.php');
                         chmod($valName.'/get.php', 0777);
                     }
                 } elseif ($valType == 'both') {
-                    mkdir($valName);
-                    chmod($valName, 0777);
                     file_put_contents($valName.'/rating', 0);
                     chmod($valName.'/rating', 0777);
                     file_put_contents($valName.'/mode', 0);
@@ -39,7 +35,9 @@ if ($mode == 'init') {
                         chmod($valName.'/get.php', 0777);
                     }
                 }
-            } else {
+            }
+        } else {
+            if (!file_exists($value)) {
                 mkdir($value);
                 chmod($value, 0777);
                 file_put_contents($value.'/rating', 0);
@@ -58,11 +56,11 @@ if ($mode == 'init') {
     }
 } elseif ($mode == 'modify') {
     foreach ($sequence as $key=>$value) {
-        if (file_exists($value)) {
-            if (strpos($value, ':') !== false) {
-                $valSep = explode(':', $value);
-                $valName = $valSep[0];
-                $valDef = $valSep[1];
+        if (strpos($value, ':') !== false) {
+            $valSep = explode(':', $value);
+            $valName = $valSep[0];
+            $valDef = $valSep[1];
+            if (file_exists($valName)) {
                 if (is_numeric($valDef)) {
                     file_put_contents($valName.'/rating', $valDef);
                     chmod($valName.'/rating', 0777);
@@ -81,11 +79,13 @@ if ($mode == 'init') {
                     file_put_contents($valName.'/mode', 0);
                     chmod($valName.'/mode', 0777);
                 }
-            } else {
-                file_put_contents($valName.'/rating', 0);
-                chmod($valName.'/rating', 0777);
-                file_put_contents($valName.'/mode', 0);
-                chmod($valName.'/mode', 0777);
+            }
+        } else {
+            if (file_exists($value)) {
+                file_put_contents($value.'/rating', 0);
+                chmod($value.'/rating', 0777);
+                file_put_contents($value.'/mode', 0);
+                chmod($value.'/mode', 0777);
             }
         }
     }
